fixed a bug in fireworks mode

+++ main/LedMode/Fireworks.h
@@ -19,7 +19,7 @@ public:
 
 protected:
     uint8_t m_fadeRate = 240; // how fast to fade: 1-fast, 256-slow
-    uint16_t m_sparkRate = 240;
+    uint8_t m_sparkRate = 240;
 };
 
 } // namespace Mode
